+++ formulario/src/main/java/com/Aplicacion/App/Repository/CarritoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CarritoRepository extends JpaRepository<CarritoModel, Integer> {
+public interface CarritoRepository extends JpaRepository<CarritoModel, Long> {
 
 }
