+++ formulario/src/main/java/com/Aplicacion/App/Repository/ProductoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<ProductoModel, Integer> {
+public interface ProductoRepository extends JpaRepository<ProductoModel, Long> {
 
 }
